After adding Lawnchair to the included JavaScript for the test.html. The code now passes all of the existing QUnit tests as well.
Where I know work has to be done is in the area of the full storage. How
does Lawnchair communicate that a given adapter is full, do all of the
existing adapters do so gracefully and in the same way, and then I once
I know that, I've got to rewrite the one section of the code that deals
with trying to set() a new value when we're already full.